+++ packages/malloy/src/lang/ast/index.ts
@@ -73,7 +73,6 @@ export * from './expressions/top-by';
 export * from './expressions/unary';
 export * from './expressions/utils';
 export * from './elements/import-statement';
-export * from './query-properties/nest-reference';
 export * from './query-properties/extend';
 export * from './parameters/has-parameter';
 export * from './query-elements/anonymous-query';
@@ -93,7 +92,6 @@ export * from './query-properties/indexing';
 export * from './query-properties/joins';
 export * from './query-properties/limit';
 export * from './query-properties/measures';
-export * from './query-properties/nest-reference';
 export * from './query-properties/nest';
 export * from './query-properties/nests';
 export * from './query-properties/ordering';
@@ -117,7 +115,6 @@ export * from './types/expression-def';
 export * from './types/field-collection-member';
 export * from './types/field-space';
 export * from './types/malloy-element';
-export * from './types/nested-query';
 export * from './types/query-element';
 export * from './types/query-item';
 export * from './types/query-property';
