+++ packages/malloy/src/lang/ast/query-properties/nest.ts
@@ -23,11 +23,9 @@
 
 import * as model from '../../../model/malloy_types';
 
-import {NestedQuery} from '../types/nested-query';
 import {FieldName, FieldSpace} from '../types/field-space';
 import {MalloyElement} from '../types/malloy-element';
 import {Noteable, extendNoteMethod} from '../types/noteable';
-import {NestReference} from './nest-reference';
 import {QueryField} from '../field-space/query-space-field';
 import {opOutputStruct} from '../struct-utils';
 import {TurtleHeadedPipe} from '../elements/pipeline-desc';
@@ -216,3 +214,51 @@ export class QueryFieldAST extends QueryField {
     return def;
   }
 }
+
+import {
+  expressionIsAggregate,
+  expressionIsAnalytic,
+  expressionIsScalar,
+  TypeDesc,
+} from '../../../model';
+import {FieldReference} from '../query-items/field-references';
+
+export class NestReference
+  extends FieldReference
+  implements QueryPropertyInterface, MakeEntry
+{
+  elementType = 'nestReference';
+  forceQueryClass = QueryClass.Grouping;
+  queryRefinementStage = LegalRefinementStage.Single;
+
+  constructor(readonly name: FieldName) {
+    super([name]);
+  }
+  typecheck(type: TypeDesc) {
+    if (type.dataType !== 'turtle') {
+      let useInstead: string;
+      let kind: string;
+      if (expressionIsAnalytic(type.expressionType)) {
+        useInstead = 'a calculate';
+        kind = 'an analytic';
+      } else if (expressionIsScalar(type.expressionType)) {
+        useInstead = 'a group_by or project';
+        kind = 'a scalar';
+      } else if (expressionIsAggregate(type.expressionType)) {
+        useInstead = 'an aggregate';
+        kind = 'an aggregate';
+      } else {
+        throw new Error(`Unexpected expression type ${type} not handled here`);
+      }
+      this.log(
+        `Cannot use ${kind} field in a nest operation, did you mean to use ${useInstead} operation instead?`
+      );
+    }
+  }
+
+  queryExecute(executeFor: Executor) {
+    executeFor.resultFS.pushFields(this);
+  }
+}
+
+export type NestedQuery = NestReference | NestDefinition | NestRefinement;
