+++ packages/malloy/src/lang/ast/query-properties/nests.ts
@@ -22,12 +22,12 @@
  */
 
 import {DefinitionList} from '../types/definition-list';
-import {NestedQuery} from '../types/nested-query';
 import {
   QueryClass,
   LegalRefinementStage,
   QueryPropertyInterface,
 } from '../types/query-property-interface';
+import {NestedQuery} from './nest';
 
 export class Nests
   extends DefinitionList<NestedQuery>
